requirements.txt
aiohttp
aiosignal
annotated-types
anyio
async-timeout
attrs
boto3
botocore
certifi
discord.py
distro
exceptiongroup
frozenlist
h11
httpcore
httpx
idna
iniconfig
jmespath
multidict
openai
packaging
pluggy
pydantic
pydantic_core
PyMySQL
pytest
pytest-asyncio
pytest-order
python-dateutil
python-dotenv
s3transfer
six
sniffio
SQLAlchemy
tomli
tqdm
typing_extensions
urllib3
yarl
